refactor: ApprovalStatus global로 변경
