+++ src/main/java/com/theocean/fundering/domain/celebrity/repository/CelebRepositoryImpl.java
@@ -3,9 +3,9 @@ package com.theocean.fundering.domain.celebrity.repository;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.theocean.fundering.domain.celebrity.domain.constant.ApprovalStatus;
 import com.theocean.fundering.domain.celebrity.dto.CelebFundingResponseDTO;
 import com.theocean.fundering.domain.celebrity.dto.CelebListResponseDTO;
+import com.theocean.fundering.global.utils.ApprovalStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
